// src/api/generated/models/commentVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UserBriefVO } from './userBriefVO';

export interface CommentVO {
  id?: string;
  appId?: string;
  parentId?: string;
  rootId?: string;
  author?: UserBriefVO;
  replyToUser?: UserBriefVO;
  content?: string;
  likeCount?: string;
  replyCount?: string;
  liked?: boolean;
  createdAt?: string;
}